switch remaining textile test files to markdown
Textile support was removed from jekyll core in #3319, and most of the
tests switched to markdown at that time.  This changes the remaining
tests to use markdown as well.  The vast majority of the test cases were
testing things in the file name or front matter, so it doesn't really
matter what markup format they use.  The one test that was claiming to
test that textile was transformed had actually been moved to markdown as
well, just not renamed.

Fixes #3507