Clean up destination modified check
Clean up the destination modified check in `source_modified_or_dest_missing?` to be easier to read. Note that it can now return `nil` instead of `false` for an unmodified `source_path` and a `nil` `dest_path`, but in a discussion on 706007ead9e97afbd0603cb8e4519b64dd325178 we decided that was okay.