docs/quickstart: update instructions to require bundler
With gem-based themes being bundled in the new site via the `Gemfile`, we
should ask folks to use Bundler wherever possible. This should lead to more
successful installations and getting the base site setup properly.

The only trouble this introduces is it puts a dependency on Bundler. That
said, I'm totally fine with requiring everyone use Bundler for this site.
How could we best install bundler in these instructions?

/cc @jekyll/documentation